requirements.txt
Flask>=2.2
PyYAML
